msm: vidc: Add validity check for video device
In a rare stability issue, it is observed that the V4L2
core has referenced a NULL video device while handling the
event from video driver. The event handler is registered
with V4L2 during set up while the video instance is opened.

Add a NULL check to ensure valid device before registering
the it with V4L2 core.

Change-Id: I1f47a433560a8f0c2b35c24e9a683dcfadae0237
Signed-off-by: Vikash Garodia <vgarodia@codeaurora.org>